+++ contrib/kube-prometheus/hack/scripts/generate-rules-configmap.sh
@@ -11,8 +11,8 @@ metadata:
 data:
 EOF
 
-for f in assets/prometheus/rules/*.rules.yaml
+for f in assets/prometheus/rules/*.rules.y*ml
 do
-  echo "  $(basename $f): |+"
+  echo "  $(basename "$f"): |+"
   cat $f | sed "s/^/    /g"
 done