+++ pkg/prometheus/promcfg_test.go
@@ -18,10 +18,10 @@ import (
 	"bytes"
 	"testing"
 
+	"k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
-	"k8s.io/client-go/pkg/api/v1"
 
 	monitoringv1 "github.com/coreos/prometheus-operator/pkg/client/monitoring/v1"
 )
@@ -63,6 +63,10 @@ func generateTestConfig(version string) ([]byte, error) {
 							Port:      intstr.FromString("web"),
 						},
 					},
+				},
+				ExternalLabels: map[string]string{
+					"label1": "value1",
+					"label2": "value2",
 				},
 				Version:  version,
 				Replicas: &replicas,