+++ pkg/prometheus/statefulset_test.go
@@ -15,14 +15,13 @@
 package prometheus
 
 import (
-	"reflect"
-	"testing"
-
 	monitoringv1 "github.com/coreos/prometheus-operator/pkg/client/monitoring/v1"
 	"github.com/stretchr/testify/require"
+	"k8s.io/api/apps/v1beta1"
+	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/pkg/api/v1"
-	"k8s.io/client-go/pkg/apis/apps/v1beta1"
+	"reflect"
+	"testing"
 )
 
 var (